remove quantum program. move its useful tests
